죽음관련 끝 몬스터 y축관련 수정 인벤 숫자표시 오른쪽정렬 몬스터 타겟이있을시 타겟을 쫒아감.

+++ PortfolioGame/PortfolioGame/Components/MapObject/Monster/MonsterUtility/monster_frame_manager.cpp
@@ -16,7 +16,7 @@ uint64_t MonsterFrameManager::GetFrameTick() const
     return _frame_tick;
 }
 
-uint64_t MonsterFrameManager::GetFrameNumber() const
+uint16_t MonsterFrameManager::GetFrameNumber() const
 {
     return _frame_number;
 }
